requirements.txt
streamlit==1.22.0
requests==2.28.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
pandas==1.5.3
Pillow==9.5.0
moviepy==1.0.3
pydub==0.25.1